// src/types.d.ts
export type TPokemonType =
  | 'bug'
  | 'dark'
  | 'dragon'
  | 'electric'
  | 'fairy'
  | 'fighting'
  | 'fire'
  | 'flying'
  | 'ghost'
  | 'grass'
  | 'ground'
  | 'ice'
  | 'normal'
  | 'poison'
  | 'psychic'
  | 'rock'
  | 'steel'
  | 'water';

export interface IPokemon {
  id: number;
  name: string;
  height: number;
  weight: number;
  sprite: string;
  types: TPokemonType[];
}